+++ Server/Models/Monsters/CastleLord.cs
@@ -15,7 +15,7 @@ namespace Server.Models.Monsters
     {
         public ConquestWar War;
 
-        public override int Attacked(MapObject attacker, int power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
+        public override int Attacked(MapObject attacker, long power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
         {
             if (attacker == null || attacker.Race != ObjectType.Player) return 0;
