+++ Server/Models/Monsters/Guard.cs
@@ -44,7 +44,7 @@ namespace Server.Models.Monsters
             NameColour = Color.SkyBlue;
         }
 
-        public override int Attacked(MapObject ob, int power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
+        public override int Attacked(MapObject ob, long power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
         {
             return 0;
         }
